refactor(map): Renamed `map` to `tiled_map` to avoid redefining the built-in
